+++ src/main/java/blackjack/domain/hand/Score.java
@@ -1,4 +1,4 @@
-package blackjack.domain.participant;
+package blackjack.domain.hand;
 
 import java.util.Objects;
