+++ src/main/java/blackjack/domain/player/PlayerNames.java
@@ -1,4 +1,4 @@
-package blackjack.domain.participant;
+package blackjack.domain.player;
 
 import java.util.List;
