+++ src/test/java/blackjack/domain/card/CardRankTest.java
@@ -2,7 +2,7 @@ package blackjack.domain.card;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import blackjack.domain.participant.Score;
+import blackjack.domain.hand.Score;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
