+++ src/test/java/blackjack/domain/dealer/DealerTest.java
@@ -1,9 +1,10 @@
-package blackjack.domain.participant;
+package blackjack.domain.dealer;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import blackjack.domain.card.Card;
 import blackjack.domain.card.CardDeck;
+import blackjack.domain.hand.Hand;
 import fixture.CardFixture;
 import fixture.HandFixture;
 import java.util.List;
