+++ src/test/java/blackjack/domain/player/PlayerNamesTest.java
@@ -1,4 +1,4 @@
-package blackjack.domain.participant;
+package blackjack.domain.player;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
