+++ src/test/java/blackjack/domain/player/ProfitTest.java
@@ -1,4 +1,4 @@
-package blackjack.domain.profit;
+package blackjack.domain.player;
 
 import static org.assertj.core.api.Assertions.assertThat;
