+++ src/test/java/fixture/DealerFixture.java
@@ -3,8 +3,8 @@ package fixture;
 import blackjack.domain.card.Card;
 import blackjack.domain.card.CardRank;
 import blackjack.domain.card.CardSuit;
-import blackjack.domain.participant.Hand;
-import blackjack.domain.participant.Dealer;
+import blackjack.domain.hand.Hand;
+import blackjack.domain.dealer.Dealer;
 import java.util.ArrayList;
 import java.util.List;
